pyproject.toml
[project]
name = "jurisbot-chile"
version = "3.0.0"
description = "Aplicación de estudio del Derecho chileno con IA"
requires-python = ">=3.10"
dependencies = [
    # === Ingesta de documentos ===
    "PyMuPDF>=1.24.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "chardet>=5.2.0",
    "ftfy>=6.2.0",
    "striprtf>=0.0.26",
    "python-magic>=0.4.27",
    "beautifulsoup4>=4.12.0",
    "Pillow>=10.0.0",
    # === NLP ===
    "spacy>=3.7.0",
    "sentence-transformers>=3.0.0",
    # === Vector DB & Search ===
    "chromadb>=0.5.0",
    # === LLM Providers ===
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    # === Base de datos ===
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # === Frontend ===
    "streamlit>=1.40.0",
    # === Utilidades ===
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
]
ocr = [
    "pytesseract>=0.3.13",
    "opencv-python-headless>=4.10.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
